[131661] RESTify: Posting via v1 controller
