+++ test/web/controller/question_test.exs
@@ -24,7 +24,6 @@ defmodule MasakiStackoverflow.Controller.QuestionTest do
       }
     ]
   }
-  @form_input_map %{title: "title987", author: "author765", body: "body432"}
 
   test "index should render every item as HTML" do
     :meck.expect(Sazabi.G2gClient, :send, fn _conn, _app_id, %{} ->
@@ -41,20 +40,5 @@ defmodule MasakiStackoverflow.Controller.QuestionTest do
         assert String.contains?(body, value)
       end)
     end)
-  end
-
-  test "create should fail if invalid values were sent" do
-    :meck.expect(Sazabi.G2gClient, :send, fn _, _, _ -> assert false end)
-    assert Req.post_form("/question", [title: "",      author: "author", body: "body"]).status == 302
-    assert Req.post_form("/question", [title: "title", author: "",       body: "body"]).status == 302
-    assert Req.post_form("/question", [title: "title", author: "author", body: ""    ]).status == 302
-  end
-
-  test "create should succeed if valid values were sent" do
-    :meck.expect(Sazabi.G2gClient, :send, fn _conn, _app_id, request ->
-      assert %Dodai.CreateDedicatedDataEntityRequest{body: %Dodai.CreateDedicatedDataEntityRequestBody{data: @form_input_map}} = request
-      Dodai.CreateDedicatedDataEntitySuccess.new(201, [], "")
-    end)
-    assert Req.post_form("/question", Map.to_list(@form_input_map)).status == 302
   end
 end
