+++ web/router.ex
@@ -1,6 +1,6 @@
 defmodule MasakiStackoverflow.Router do
   use SolomonLib.Router
 
-  get  "/question", Question, :index
-  post "/question", Question, :create
+  get  "/question",    Question,    :index
+  post "/v1/question", V1.Question, :create
 end
